backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
django-filter>=23.5
djangorestframework-simplejwt>=5.3.0 